+++ appGUI/preferences/geometry/GeometryAdvOptPrefGroupUI.py
@@ -34,6 +34,8 @@ class GeometryAdvOptPrefGroupUI(OptionsGroupUI):
         self.layout.addWidget(self.geo_label)
 
         grid1 = FCGridLayout(v_spacing=5, h_spacing=3)
+        grid1.setColumnStretch(0, 0)
+        grid1.setColumnStretch(1, 1)
         self.layout.addLayout(grid1)
 
         # Size of trace segment on X axis